Extract mesh support ResourcePage planner
Move mesh support ResourcePage planning out of the large core weave facade into a dedicated module while preserving existing exports from core/weave/weave.ts.

Add small shared modules for WeaveInputError and VersionPlan so extracted planners can avoid importing through the facade. Document the refactor slice and current fixture-ladder test status.
