+++ src/core/weave/errors.ts
@@ -1,0 +1,6 @@
+export class WeaveInputError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = "WeaveInputError";
+  }
+}
